+++ readme/CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Version 1.0 (upcoming)
+* Added: The ability to check if the provided url is probable to contain script(s) vulnerable to shellshock.
 * Revised: Multiple eye-candy revisions have been performed.
 * Fixed: HTTPS requests fixation, if the "--proxy" option is enabled.
 * Fixed: Multiple fixes regarding the shellshock module have been performed.